Refactor ImageUploader into a separate component
