+++ src/components/App/App.css
@@ -22,7 +22,7 @@
 }
 
 .add-post-button:hover {
-  background-color: #d3ab08;
+  background-color: #dfb509;
 }
 
 .add-post-button:active {
