+++ src/components/ImageUploader/ImageUploader.css
@@ -1,27 +1,3 @@
-.modal-wrapper {
-  position: fixed;
-  z-index: 1;
-  padding-top: 7em;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgba(0, 0, 0, 0.5);
-  text-align: center;
-}
-
-.close-modal {
-  align-self: flex-start;
-  border: 1px solid black;
-  padding: 0.25em;
-  margin-top: 1em;
-}
-ß .close-modal:hover {
-  background-color: whitesmoke;
-  transform: scale(1.1);
-}
-
 .modal-content {
   margin: auto;
   background-color: white;
@@ -39,6 +15,20 @@
   font-weight: bolder;
 }
 
+.post-prompt {
+  font-size: large;
+}
+
+.image-item {
+  margin-top: 1em;
+}
+
+.post-description {
+  width: 100%;
+  text-align: start;
+  padding-bottom: 3em;
+}
+
 .post-submit-btn {
   margin: 1em;
 }
@@ -50,6 +40,20 @@
 
 button {
   cursor: pointer;
+}
+
+.hidden {
+  display: none;
+}
+
+.error {
+  color: red;
+  font-size: 70%;
+}
+
+input {
+  word-wrap: break-word;
+  word-break: break-all;
 }
 
 @media screen and (min-width: 768px) {
